backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.36
bcrypt==4.2.1
PyJWT==2.9.0
# ML stack: use flexible pins so pip can pick wheels for your Python (3.11–3.13 recommended; 3.14 may lag).
pandas>=2.0,<3
numpy>=1.26,<3
scikit-learn>=1.4,<2
imbalanced-learn>=0.12,<1
xgboost>=2.0,<3
shap>=0.44,<1
joblib>=1.3,<2
